Commit 06/11
Buscando error en profesorid

+++ src/main/java/ar/edu/utn/frbb/tup/business/impl/ProfesorServiceImpl.java
@@ -15,7 +15,7 @@ public class ProfesorServiceImpl implements ProfesorService {
     private ProfesorDao dao;
 
     @Override
-    public Profesor buscarProfesor(long id) {
+    public Profesor buscarProfesor(int id) {
         return dao.get(id);
     }
 }
