+++ src/main/java/ar/edu/utn/frbb/tup/model/dto/MateriaDto.java
@@ -9,13 +9,13 @@ public class MateriaDto {
     private String nombre;
     private int anio;
     private int cuatrimestre;
-    private long profesorId;
+    private int profesorId;
 
-    public long getProfesorId() {
+    public int getProfesorId() {
         return profesorId;
     }
 
-    public void setProfesorId(long profesorId) {
+    public void setProfesorId(int profesorId) {
         this.profesorId = profesorId;
     }
