+++ src/main/java/ar/edu/utn/frbb/tup/persistence/ProfesorDao.java
@@ -3,5 +3,5 @@ package ar.edu.utn.frbb.tup.persistence;
 import ar.edu.utn.frbb.tup.model.Profesor;
 
 public interface ProfesorDao {
-    public Profesor get(long id);
+    public Profesor get(int id);
 }
